Move creatable saved query types to /models
