+++ src/functions/saved-queries/get-one-saved-query.spec.ts
@@ -6,10 +6,10 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
-import { isSavedQuery, SavedQuery } from '../../models';
+import { CreatableSavedQuery, isSavedQuery, SavedQuery } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
-import { CreatableSavedQuery, makeCreateOneSavedQuery } from './create-one-saved-query';
+import { makeCreateOneSavedQuery } from './create-one-saved-query';
 import { makeDeleteOneSavedQuery } from './delete-one-saved-query';
 import { makeGetAllSavedQueries } from './get-all-saved-queries';
 import { makeGetOneSavedQuery } from './get-one-saved-query';
