+++ src/functions/saved-queries/get-saved-queries-authorized-to-me.spec.ts
@@ -7,12 +7,12 @@
  **************************************************************************/
 
 import { random, sortBy } from 'lodash';
-import { isSavedQuery, SavedQuery, User } from '../../models';
+import { CreatableSavedQuery, isSavedQuery, SavedQuery, User } from '../../models';
 import { integrationTest } from '../../tests';
 import { TEST_AUTH_TOKEN, TEST_HOST } from '../../tests/config';
 import { makeLoginOneUser } from '../auth/login-one-user';
 import { CreatableUser, makeCreateOneUser, makeGetOneUser } from '../users';
-import { CreatableSavedQuery, makeCreateOneSavedQuery } from './create-one-saved-query';
+import { makeCreateOneSavedQuery } from './create-one-saved-query';
 import { makeDeleteOneSavedQuery } from './delete-one-saved-query';
 import { makeGetAllSavedQueries } from './get-all-saved-queries';
 import { makeGetSavedQueriesAuthorizedToMe } from './get-saved-queries-authorized-to-me';
